+++ sentry-opentelemetry/sentry-opentelemetry-core/src/main/java/io/sentry/opentelemetry/SentrySpanExporter.java
@@ -10,6 +10,7 @@ import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.trace.data.SpanData;
 import io.opentelemetry.sdk.trace.data.StatusData;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
+import io.opentelemetry.semconv.ResourceAttributes;
 import io.opentelemetry.semconv.SemanticAttributes;
 import io.sentry.Baggage;
 import io.sentry.DateUtils;
@@ -64,7 +65,9 @@ public final class SentrySpanExporter implements SpanExporter {
           InternalSemanticAttributes.SAMPLE_RATE.getKey(),
           InternalSemanticAttributes.PROFILE_SAMPLED.getKey(),
           InternalSemanticAttributes.PROFILE_SAMPLE_RATE.getKey(),
-          InternalSemanticAttributes.PARENT_SAMPLED.getKey());
+          InternalSemanticAttributes.PARENT_SAMPLED.getKey(),
+          ResourceAttributes.PROCESS_COMMAND_ARGS.getKey() // can be very long
+          );
   private static final @NotNull Long SPAN_TIMEOUT = DateUtils.secondsToNanos(5 * 60);
 
   public static final String TRACE_ORIGIN = "auto.opentelemetry";
@@ -483,7 +486,7 @@ public final class SentrySpanExporter implements SpanExporter {
           (key, value) -> {
             if (key != null) {
               final @NotNull String stringKey = key.getKey();
-              if (!isSentryInternalKey(stringKey)) {
+              if (!shouldRemoveAttribute(stringKey)) {
                 mapWithStringKeys.put(stringKey, value);
               }
             }
@@ -493,7 +496,7 @@ public final class SentrySpanExporter implements SpanExporter {
     return mapWithStringKeys;
   }
 
-  private boolean isSentryInternalKey(final @NotNull String key) {
+  private boolean shouldRemoveAttribute(final @NotNull String key) {
     return attributeKeysToRemove.contains(key);
   }